requirements.txt
streamlit
openai
python-dotenv
pandas
pdfplumber
python-docx
python-pptx
faiss-cpu
Chroma
